* Added Test Summary to summarize test failures and causes at the very end of stdout
* Refactored Logre-light to output detailed logs after all tests have completed (so that now one test failure doesn't cause a bunch of test time outs)
* Fixed one remaining test that was still sporadically failing
* Added precomputation for genData to drastically improve its performance

Testing Done:
Manual Testing by negating test asserts and checking output, and then restoring test and ensuring everything passes.

Reviewed at http://us118/review/r/190/

+++ daemon/TestLibrary_test.go
@@ -34,6 +34,15 @@ const fusectlPath = "/sys/fs/fuse/"
 
 type quantumFsTest func(test *testHelper)
 
+type logscanError struct {
+	logFile           string
+	shouldFailLogscan bool
+	testName          string
+}
+
+var errorMutex sync.Mutex
+var errorLogs []logscanError
+
 // startTest is a helper which configures the testing environment
 func runTest(t *testing.T, test quantumFsTest) {
 	t.Parallel()
@@ -51,6 +60,7 @@ func runTest(t *testing.T, test quantumFsTest) {
 		cachePath:  cachePath,
 		logger:     qlog.NewQlogExt(cachePath+"/ramfs", 60*10000*24),
 	}
+	th.createTestDirs()
 
 	defer th.endTest()
 
@@ -62,7 +72,7 @@ func runTest(t *testing.T, test quantumFsTest) {
 
 	select {
 	case <-time.After(1000 * time.Millisecond):
-		testResult = "TIMED OUT"
+		testResult = "ERROR: TIMED OUT"
 
 	case testResult = <-th.testResult:
 	}
@@ -153,11 +163,11 @@ func (th *testHelper) endTest() {
 			runtime.GC()
 
 			if err := th.qfs.server.Unmount(); err != nil {
-				th.t.Fatalf("Failed to unmount quantumfs instance "+
-					"after aborting: %v", err)
+				th.t.Fatalf("ERROR: Failed to unmount quantumfs "+
+					"instance after aborting: %v", err)
 			}
-			th.t.Fatalf("Failed to unmount quantumfs instance, are you"+
-				" leaking a file descriptor?: %v", err)
+			th.t.Fatalf("ERROR: Failed to unmount quantumfs instance, "+
+				"are you leaking a file descriptor?: %v", err)
 		}
 	}
 
@@ -165,12 +175,14 @@ func (th *testHelper) endTest() {
 		th.waitToBeUnmounted()
 		time.Sleep(1 * time.Second)
 
-		th.logscan()
-
-		if err := os.RemoveAll(th.tempDir); err != nil {
-			th.t.Fatalf("Failed to cleanup temporary mount point: %v",
-				err)
+		if testFailed := th.logscan(); !testFailed {
+			if err := os.RemoveAll(th.tempDir); err != nil {
+				th.t.Fatalf("Failed to cleanup temporary mount "+
+					"point: %v", err)
+			}
 		}
+	} else {
+		th.t.Fatalf("No temporary directory available for logs")
 	}
 
 	if exception != nil {
@@ -196,35 +208,68 @@ func (th *testHelper) waitToBeUnmounted() {
 }
 
 // Check the test output for errors
-func (th *testHelper) logscan() {
+func (th *testHelper) logscan() (foundErrors bool) {
 	// Check the format string map for the log first to speed this up
-	logFile := th.qfs.config.CachePath + "/qlog"
-	if !qlog.LogscanSkim(logFile) {
-		return
+	logFile := th.tempDir + "/ramfs/qlog"
+	errorsPresent := qlog.LogscanSkim(logFile)
+
+	// Nothing went wrong if either we should fail and there were errors,
+	// or we shouldn't fail and there weren't errors
+	if th.shouldFailLogscan == errorsPresent {
+		return false
 	}
 
+	// There was a problem
+	errorMutex.Lock()
+	errorLogs = append(errorLogs, logscanError{
+		logFile:           logFile,
+		shouldFailLogscan: th.shouldFailLogscan,
+		testName:          th.testName,
+	})
+	errorMutex.Unlock()
+
+	if !th.shouldFailLogscan {
+		th.t.Fatalf("Test FAILED due to FATAL messages\n")
+	} else {
+		th.t.Fatalf("Test FAILED due to missing FATAL messages\n")
+	}
+
+	return true
+}
+
+func outputLogError(errInfo logscanError) (summary string) {
 	errors := make([]string, 0, 10)
-	testOutput := qlog.ParseLogs(logFile)
+	testOutput := qlog.ParseLogs(errInfo.logFile)
 
 	lines := strings.Split(testOutput, "\n")
+
+	extraLines := 0
 	for _, line := range lines {
 		if strings.Contains(line, "PANIC") ||
 			strings.Contains(line, "WARN") ||
 			strings.Contains(line, "ERROR") {
+			extraLines = 2
+		}
+
+		// Output a couple extra lines after an ERROR
+		if extraLines > 0 {
 			errors = append(errors, line)
+			extraLines--
 		}
 	}
 
-	if !th.shouldFailLogscan && len(errors) != 0 {
-		for _, err := range errors {
-			th.t.Logf("FATAL message logged: %s", err)
-		}
-		th.t.Fatalf("Test FAILED due to FATAL messages. Dumping Logs:\n%s\n"+
-			"--- Test %s FAILED\n\n", testOutput, th.testName)
-	} else if th.shouldFailLogscan && len(errors) == 0 {
-		th.t.Fatalf("Test FAILED due to missing FATAL messages."+
-			" Dumping Logs:\n%s\n--- Test %s FAILED\n\n", testOutput,
-			th.testName)
+	if !errInfo.shouldFailLogscan {
+		fmt.Printf("Test %s FAILED due to ERROR. Dumping Logs:\n%s\n"+
+			"--- Test %s FAILED\n\n\n", errInfo.testName, testOutput,
+			errInfo.testName)
+		return fmt.Sprintf("--- Test %s FAILED due to errors:\n%s\n",
+			errInfo.testName, strings.Join(errors, "\n"))
+	} else {
+		fmt.Printf("Test %s FAILED due to missing FATAL messages."+
+			" Dumping Logs:\n%s\n--- Test %s FAILED\n\n\n",
+			errInfo.testName, testOutput, errInfo.testName)
+		return fmt.Sprintf("--- Test %s FAILED\nExpected errors, but found"+
+			" none.\n", errInfo.testName)
 	}
 }
 
@@ -245,20 +290,18 @@ type testHelper struct {
 	shouldFailLogscan bool
 }
 
-func (th *testHelper) createTestDirs() string {
+func (th *testHelper) createTestDirs() {
 	th.tempDir = testRunDir + "/" + th.testName
-	mountPath := th.tempDir + "/mnt"
 
+	mountPath := th.tempDir + "/mnt"
 	os.MkdirAll(mountPath, 0777)
 	th.log("Using mountpath %s", mountPath)
 
 	os.MkdirAll(th.tempDir+"/ether", 0777)
-
-	return mountPath
 }
 
 func (th *testHelper) defaultConfig() QuantumFsConfig {
-	mountPath := th.createTestDirs()
+	mountPath := th.tempDir + "/mnt"
 
 	config := QuantumFsConfig{
 		CachePath:        th.tempDir + "/ramfs",
@@ -535,7 +578,21 @@ func TestMain(m *testing.M) {
 	// will never return to allow GC to progress, the test program is deadlocked.
 	debug.SetGCPercent(-1)
 
+	// Precompute a bunch of our genData to save time during tests
+	genData(40 * 1024 * 1024)
+
+	// Setup an array for tests with errors to be logscanned later
+	errorLogs = make([]logscanError, 0)
+
 	result := m.Run()
+
+	testSummary := ""
+	errorMutex.Lock()
+	for i := 0; i < len(errorLogs); i++ {
+		testSummary += outputLogError(errorLogs[i])
+	}
+	errorMutex.Unlock()
+	fmt.Println("------ Test Summary:\n" + testSummary)
 
 	os.RemoveAll(testRunDir)
 	os.Exit(result)
@@ -738,4 +795,37 @@ func (test *testHelper) fileSize(filename string) int64 {
 	err := syscall.Stat(filename, &stat)
 	test.assert(err == nil, "Error stat'ing test file: %v", err)
 	return stat.Size
+}
+
+var genDataMutex sync.RWMutex
+var precompGenData []byte
+var genDataLast int
+
+func genData(maxLen int) []byte {
+	if maxLen > len(precompGenData) {
+		// we need to expand the array
+		genDataMutex.Lock()
+
+		for len(precompGenData) <= maxLen {
+			precompGenData = append(precompGenData,
+				strconv.Itoa(genDataLast)...)
+			genDataLast++
+		}
+
+		genDataMutex.Unlock()
+	}
+	genDataMutex.RLock()
+	defer genDataMutex.RUnlock()
+
+	return precompGenData[:maxLen]
+}
+
+func TestGenData_test(t *testing.T) {
+	runTest(t, func(test *testHelper) {
+		hardcoded := "012345678910111213141516171819202122232425262"
+		data := genData(len(hardcoded))
+
+		test.assert(bytes.Equal([]byte(hardcoded), data),
+			"Data gen function off: %s vs %s", hardcoded, data)
+	})
 }
